backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-Bcrypt==1.0.1
Twilio==9.5.0
python-dotenv==1.0.1
werkzeug>=3.1.3
flask-jwt-extended>=4.0.0
flask-cors>=4.0.0